+++ R/RLCS_train_SL_specific.R
@@ -126,6 +126,7 @@
   t_labs <- sapply(pop, \(x) { x$action })
 
   pop_size <- length(pop)
+
   subsumers_list <- lapply(1:(pop_size-1), \(item) {
 
     # if(pop[[item]]$numerosity > 0) { ## Now superfluous in vectorized approach
@@ -145,6 +146,7 @@
 
     rest_t_matrices_zeros <- t_matrices[[1]][(item+1):pop_size,]
     rest_t_matrices_ones <- t_matrices[[2]][(item+1):pop_size,]
+
     ## Remember the rules are sorted by accuracy and generality already!
     subsumer_must_match_zeros <- rest_t_matrices_zeros  %*% t_matrices[[1]][item,]
     subsumer_must_match_ones <- rest_t_matrices_ones  %*% t_matrices[[2]][item,]
@@ -198,79 +200,94 @@
 }
 
 
-.apply_subsumption_whole_pop_sl_env <- function(env) {
-
-
-  browser()
-  if(length(env$lcs$pop) > 1) {
-    # pop <- .lcs_best_sort_sl(pop)
-    .lcs_best_sort_sl_env(env)
-    .apply_deletion_no_threshold_env(env)
-
-    ## Within this function in this case :)
-    t_matrices <- .recalculate_pop_matrices(env$lcs$pop)
-
-    # browser()
-    t_labs <- sapply(env$lcs$pop, \(x) { x$action })
-    subsumers_list <- lapply(1:(length(env$lcs$pop)-1), \(item) {
-
-      # if(pop[[item]]$numerosity > 0) { ## Now superfluous in vectorized approach
-
-      t_zero <- env$lcs$pop[[item]]$condition_list$"0"
-      t_one <- env$lcs$pop[[item]]$condition_list$"1"
-      # cond_string <- pop[[item]]$condition_string
-      # cond_lab <- pop[[item]]$action
-
-      pop_to_delete <- NULL
-
-      # # ## Alternative approach now, let's see!
-
-
-      ## Remember the rules are sorted by accuracy and generality already!
-      subsumer_must_match_zeros <- t_matrices[[1]][(item+1):length(env$lcs$pop),]  %*% t_matrices[[1]][item,]
-      subsumer_must_match_ones <- t_matrices[[2]][(item+1):length(env$lcs$pop),]  %*% t_matrices[[2]][item,]
-
-      subsumed_must_be_different_zero <-  t_matrices[[1]][(item+1):length(env$lcs$pop),] %*% t_matrices[[2]][item,]
-      subsumed_must_be_different_one <- t_matrices[[2]][(item+1):length(env$lcs$pop),] %*% t_matrices[[1]][item,]
-
-      pop_to_delete <- which(
-        ((subsumer_must_match_zeros+subsumer_must_match_ones) == (length(t_zero)+length(t_one))) &
-          (subsumed_must_be_different_zero+subsumed_must_be_different_one == 0) &
-          (t_labs[(item+1):length(env$lcs$pop)] == t_labs[item])
-      )
-
-      if(!is.null(pop_to_delete) && length(pop_to_delete) > 0) {
-        pop_to_delete <- pop_to_delete + item ## Optimization. POSITIONS
-        return(pop_to_delete)
-      }
-
-      return(c()) ## Default: Nothing to delete
-    })
-
-    ## Update numerosity for each subsumer
-    ## Before vectorization
-    # for(i in 1:length(subsumers_list)) {
-    #   pop[[i]]$numerosity <- pop[[i]]$numerosity+length(subsumers_list[[i]])
-    # } ## Probably can be vectorized, this too...
-    ## Vectorized version
-    subsumers_positions <- which(sapply(subsumers_list, \(x) !is.null(x)))
-    pop[subsumers_positions] <- lapply(subsumers_positions, \(i) {
-      env$lcs$pop[[i]]$numerosity <- env$lcs$pop[[i]]$numerosity + length(subsumers_list[[i]])
-      env$lcs$pop[[i]]
-    }) ## Go this just once!
-
-    pop_to_delete <- unique(unlist(subsumers_list)) ## Reduce operations
-    env$lcs$pop[pop_to_delete] <- lapply(env$lcs$pop[pop_to_delete], \(item_to_clean) {
-      item_to_clean$numerosity <- 0
-      item_to_clean
-    }) ## Go this just once!
-
-    env$lcs$pop <- .apply_deletion_no_threshold(env$lcs$pop)
-  }
-
-  # pop
-}
-
+# .apply_subsumption_whole_pop_sl_env <- function(env) {
+#
+#
+#   browser()
+#   if(length(env$lcs$pop) > 1) {
+#     # pop <- .lcs_best_sort_sl(pop)
+#     .lcs_best_sort_sl_env(env)
+#     .apply_deletion_no_threshold_env(env)
+#
+#     ## Within this function in this case :)
+#     t_matrices <- .recalculate_pop_matrices(env$lcs$pop)
+#
+#     # browser()
+#     t_labs <- sapply(env$lcs$pop, \(x) { x$action })
+#
+#
+#
+#
+#     subsumers_list <- lapply(1:(length(env$lcs$pop)-1), \(item) {
+#
+#       # if(pop[[item]]$numerosity > 0) { ## Now superfluous in vectorized approach
+#
+#       t_zero <- env$lcs$pop[[item]]$condition_list$"0"
+#       t_one <- env$lcs$pop[[item]]$condition_list$"1"
+#       # cond_string <- pop[[item]]$condition_string
+#       # cond_lab <- pop[[item]]$action
+#
+#       pop_to_delete <- NULL
+#
+#       # # ## Alternative approach now, let's see!
+#
+#
+#       ## Remember the rules are sorted by accuracy and generality already!
+#       if(env$use_gpu) {
+#         # message("using gpu")
+#         subsumer_must_match_zeros <- as.vector(as.matrix(torch::torch_matmul(t_matrices[[1]][(item+1):length(env$lcs$pop),], t_matrices[[1]][item,])))
+#         subsumer_must_match_ones <- as.vector(as.matrix(torch::torch_matmul(t_matrices[[2]][(item+1):length(env$lcs$pop),], t_matrices[[2]][item,])))
+#
+#         subsumed_must_be_different_zero <-  as.vector(as.matrix(torch::torch_matmul(t_matrices[[1]][(item+1):length(env$lcs$pop),], t_matrices[[2]][item,])))
+#         subsumed_must_be_different_one <- as.vector(as.matrix(torch::torch_matmul(t_matrices[[2]][(item+1):length(env$lcs$pop),], t_matrices[[1]][item,])))
+#
+#       } else {
+#         subsumer_must_match_zeros <- t_matrices[[1]][(item+1):length(env$lcs$pop),]  %*% t_matrices[[1]][item,]
+#         subsumer_must_match_ones <- t_matrices[[2]][(item+1):length(env$lcs$pop),]  %*% t_matrices[[2]][item,]
+#
+#         subsumed_must_be_different_zero <-  t_matrices[[1]][(item+1):length(env$lcs$pop),] %*% t_matrices[[2]][item,]
+#         subsumed_must_be_different_one <- t_matrices[[2]][(item+1):length(env$lcs$pop),] %*% t_matrices[[1]][item,]
+#
+#       }
+#
+#       pop_to_delete <- which(
+#         ((subsumer_must_match_zeros+subsumer_must_match_ones) == (length(t_zero)+length(t_one))) &
+#           (subsumed_must_be_different_zero+subsumed_must_be_different_one == 0) &
+#           (t_labs[(item+1):length(env$lcs$pop)] == t_labs[item])
+#       )
+#
+#       if(!is.null(pop_to_delete) && length(pop_to_delete) > 0) {
+#         pop_to_delete <- pop_to_delete + item ## Optimization. POSITIONS
+#         return(pop_to_delete)
+#       }
+#
+#       return(c()) ## Default: Nothing to delete
+#     })
+#
+#     ## Update numerosity for each subsumer
+#     ## Before vectorization
+#     # for(i in 1:length(subsumers_list)) {
+#     #   pop[[i]]$numerosity <- pop[[i]]$numerosity+length(subsumers_list[[i]])
+#     # } ## Probably can be vectorized, this too...
+#     ## Vectorized version
+#     subsumers_positions <- which(sapply(subsumers_list, \(x) !is.null(x)))
+#     pop[subsumers_positions] <- lapply(subsumers_positions, \(i) {
+#       env$lcs$pop[[i]]$numerosity <- env$lcs$pop[[i]]$numerosity + length(subsumers_list[[i]])
+#       env$lcs$pop[[i]]
+#     }) ## Go this just once!
+#
+#     pop_to_delete <- unique(unlist(subsumers_list)) ## Reduce operations
+#     env$lcs$pop[pop_to_delete] <- lapply(env$lcs$pop[pop_to_delete], \(item_to_clean) {
+#       item_to_clean$numerosity <- 0
+#       item_to_clean
+#     }) ## Go this just once!
+#
+#     env$lcs$pop <- .apply_deletion_no_threshold(env$lcs$pop)
+#   }
+#
+#   # pop
+# }
+#
 
 .apply_deletion_sl <- function(lcs, deletion_limit = 0.6, max_pop_size = 10000) {
 
@@ -295,6 +312,30 @@
   lcs
 }
 
+
+# .apply_deletion_sl_env <- function(lcs, env, deletion_limit = 0.6, max_pop_size = 10000) {
+#
+#   lcs$pop <- lapply(lcs$pop, \(x) {
+#     if(x$accuracy < deletion_limit) x$numerosity <- 0
+#     x
+#   })
+#
+#   if(length(lcs$pop) > max_pop_size) {
+#     lcs$pop <- .lcs_best_sort_sl(lcs$pop)
+#     lcs$pop[max_pop_size:length(lcs$pop)] <- lapply(lcs$pop[max_pop_size:length(lcs$pop)], \(x) {
+#       x$numerosity <- 0
+#       x
+#     })
+#   }
+#
+#   lcs$pop <- .apply_deletion_no_threshold(lcs$pop)
+#   # lcs$matrices <- .recalculate_pop_matrices(lcs$pop)
+#   lcs$matrices <- .recalculate_pop_matrices_env(lcs$pop, env)
+#   lcs$lengths <- .lengths_fixed_bits(lcs$pop)
+#   lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
+#
+#   lcs
+# }
 
 ######
 ## KEY FUNCTION: Train binary classifier LCS
@@ -394,7 +435,9 @@
     lcs <- .apply_deletion_sl(lcs,
                               deletion_limit = run_params$get_deletion_threshold(),
                               max_pop_size = run_params$get_max_pop_size())
-
+    apply_deletion_sl_env(environment(),
+                          deletion_limit = run_params$get_deletion_threshold(),
+                          max_pop_size = run_params$get_max_pop_size())
     print(paste("Epoch:", n_epoch,
                 "Progress Exposure:", train_count,
                 "Classifiers Count:", length(lcs$pop)))
@@ -443,6 +486,43 @@
   lcs
 }
 
+.perfect_coverage_simplifier_sl_env <- function(lcs, env, train_env_df, t_classes_counts) {
+  # print(as.data.frame(t_classes_counts))
+  pop_actions <- sapply(lcs$pop, \(x) x$action)
+  # print(table(pop_actions))
+  # print(length(t_classes_counts))
+  if(length(unique(pop_actions)) == length(t_classes_counts)) {
+    t_df <- data.frame(match_sizes = .reverse_match_set_size(lcs$pop, train_env_df),
+                       match_class = pop_actions,
+                       rule_id = 1:length(lcs$pop))
+    t_df <- merge(t_df, as.data.frame(t_classes_counts), by.x = "match_class", by.y = "Var1")
+    t_df <- t_df[order(t_df$rule_id),]
+    # print(head(t_df, 10))
+
+    if(nrow(t_df) > 1) {
+      for(i in 1:(nrow(t_df)-1)) {
+
+        if(t_df$match_sizes[i] == t_df$Freq[i]) { ## Full Coverage!
+          ## Then delete all entries that have the same match_class, as they are useless
+          to_remove <- which(t_df[i+1:nrow(t_df), "match_class"] == t_df$match_class[i])
+          lcs$pop[to_remove] <- lapply(lcs$pop[to_remove], \(item) {
+            item$numerosity <- 0
+            item
+          })
+        }
+      }
+      # lcs <- .apply_deletion_sl(lcs)
+      use_gpu <- env$use_gpu
+      if(use_gpu) gpu_type <- env$gpu_type
+      .apply_deletion_sl_env(environment())
+      # pop_actions <- sapply(lcs$pop, \(x) x$action)
+      # print(table(pop_actions))
+    }
+  }
+
+  lcs
+}
+
 .apply_deletion_sl_env <- function(env, deletion_limit = 0.6, max_pop_size = 10000) {
 
   env$lcs$pop <- lapply(env$lcs$pop, \(x) {
@@ -461,7 +541,8 @@
 
   #env$lcs$pop <- .apply_deletion_no_threshold(env$lcs$pop)
   .apply_deletion_no_threshold_env(env)
-  env$lcs$matrices <- .recalculate_pop_matrices(env$lcs$pop)
+  # env$lcs$matrices <- .recalculate_pop_matrices(env$lcs$pop)
+  env$lcs$matrices <- .recalculate_pop_matrices_env(env$lcs$pop, env)
   env$lcs$lengths <- .lengths_fixed_bits(env$lcs$pop)
   env$lcs$actions_vec <- .recalculate_actions_vec(env$lcs$pop)
 
@@ -479,6 +560,12 @@
   ## Main process for R LCS Training
   ######
   # list(pop = list(), matrices = list(), lengths = 0)
+
+  if(env$use_gpu) {
+    use_gpu <- env$use_gpu
+    gpu_type <- env$gpu_type
+  }
+
 
   ## ADD ERROR CONTROL
   t_instance_state <- t_instance$state
@@ -645,6 +732,9 @@
 #' @param max_pop_size_parallel
 #' Defaults to 10000. OPTIONAL. ONLY USED if foreach and doParallel are available
 #' Applies as last step: additional deletion to contain population sizes after merging.
+#' @param use_gpu
+#' Defaults to FALSE. OPTIONAL. ONLY USED if torch is available. Defaults to CUDA if found.
+#' Otherwise goes back to CPU (and is slower than not enabling the GPU option in the first place).
 #'
 #' @returns
 #' An \R \code{RLCS Model} containing the proposed model, made of several classifiers.
@@ -669,10 +759,16 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
                           merge_best_n = 0,
                           second_evolution_iterations = 1,
                           second_evolution_run_params = NULL,
-                          max_pop_size_parallel = 10000) {
+                          max_pop_size_parallel = 10000,
+                          use_gpu = F) {
   ## Initialization:
   lcs <- .new_rlcs()
   .validate_SL_train_df(train_env_df)
+  if(use_gpu & requireNamespace("torch", quietly=T)) {
+    use_gpu <- use_gpu
+    gpu_type <- ifelse(torch::cuda_is_available(), "cuda", ifelse(torch::backends_mps_is_available(), "mps", "cpu"))
+  }
+
 
   ## TODO Add Running Params Checks here...
   ## Further testing will surface issues here.
@@ -681,7 +777,9 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
   if(!is.null(pre_trained_lcs)) lcs <- pre_trained_lcs
 
   ## NEW
-  lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+  # lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+  # if(env$use_gpu) { use_gpu <- env$use_gpu; gpu_type <- env$gpu_type }
+  lcs$matrices <- .recalculate_pop_matrices_env(lcs$pop, environment())
   # lcs$lengths <- sapply(lcs$pop, \(x) x$length_fixed_bits) ## Poor naming...
   lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, numeric(1)) ## Poor naming...
   lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
@@ -716,11 +814,11 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
           for(i in 1:size_env) {
             #sub_lcs <-
             .rlcs_train_one_instance_one_epoch_mat_env(environment(),
-                                                              sub_df[i, ],
-                                                              size_env,
-                                                              epoch,
-                                                              (epoch-1)*size_env+i, ## train_count
-                                                              run_params)
+                                                       sub_df[i, ],
+                                                       size_env,
+                                                       epoch,
+                                                       (epoch-1)*size_env+i, ## train_count
+                                                       run_params)
           }
 
         }
@@ -742,7 +840,8 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
       lcs$pop <- .apply_subsumption_whole_pop_sl(lcs$pop)
       # .apply_subsumption_whole_pop_sl_env(environment())
       print(length(lcs$pop))
-      lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
+      # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
+      .apply_deletion_sl_env(environment(), max_pop_size = max_pop_size_parallel)
       # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
       print(length(lcs$pop))
       # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
@@ -850,7 +949,8 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
 
           compacted_classifier <- .apply_subsumption_whole_pop_sl(compacted_classifier)
           lcs$pop <- compacted_classifier
-          lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+          # lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+          lcs$matrices <- .recalculate_pop_matrices_env(lcs$pop, environment()) ## Poor naming...
           lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, numeric(1)) ## Poor naming...
           lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
         } else {
@@ -903,7 +1003,8 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
     ##
     ## Final simplification: Coverage
     ##
-    lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
+    # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
+    lcs <- .perfect_coverage_simplifier_sl_env(lcs, environment(), train_env_df, t_classes_counts)
   }
 
   ## Sometimes, deletion removes all rules as none are good enough!
@@ -916,235 +1017,235 @@ rlcs_train_sl <- function(train_env_df, run_params = RLCS_hyperparameters(),
 
 
 
-.rlcs_train_sl_old <- function(train_env_df, run_params = RLCS_hyperparameters(),
-                              pre_trained_lcs = NULL, verbose=FALSE,
-                              n_agents = 0,
-                              split_horizontal = F, ## That is but one option!
-                              use_validation=F,
-                              merge_best_n = 0,
-                              second_evolution_iterations = 1,
-                              second_evolution_run_params = NULL,
-                              max_pop_size_parallel = 10000) {
-  ## Initialization:
-  lcs <- .new_rlcs()
-  .validate_SL_train_df(train_env_df)
-
-  ## TODO Add Running Params Checks here...
-  ## Further testing will surface issues here.
-
-  ## Re-training, or "online" updates
-  if(!is.null(pre_trained_lcs)) lcs <- pre_trained_lcs
-
-  ## NEW
-  lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
-  # lcs$lengths <- sapply(lcs$pop, \(x) x$length_fixed_bits) ## Poor naming...
-  lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, integer(1)) ## Poor naming...
-  lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
-
-  ## Shuffling population, just in case...
-  train_env_df <- train_env_df[sample(1:nrow(train_env_df),
-                                      nrow(train_env_df),
-                                      replace = F), ]
-
-  t_classes_counts <- table(train_env_df$class) ## For Coverage!!
-
-  if(requireNamespace("foreach", quietly=T) &
-     requireNamespace("doParallel", quietly=T) &
-     n_agents > 1) { ## NEW!
-
-    if(split_horizontal) {
-
-      agents <- foreach::foreach(i = 1:n_agents) %dopar% { ## Train N agents
-
-        sets_size <- floor(nrow(train_env_df) / n_agents)
-        sub_start <- (i-1)*sets_size+1
-        sub_end <- i*sets_size
-        sub_df <- train_env_df[sub_start:sub_end,]
-        t_classes_counts <- table(sub_df$class) ## For Coverage!!
-
-        library(RLCS) ## Assuming you've gotten the package installed by now...
-
-        size_env <- nrow(sub_df)
-
-        sub_lcs <- lcs
-
-        for(epoch in 1:(run_params$get_n_epochs())) {
-          for(i in 1:size_env) {
-            sub_lcs <- .rlcs_train_one_instance_one_epoch_mat(sub_lcs,
-                                                              sub_df[i, ],
-                                                              size_env,
-                                                              epoch,
-                                                              (epoch-1)*size_env+i, ## train_count
-                                                              run_params)
-          }
-        }
-
-        ## Sometimes, deletion removes all rules as none are good enough!
-        if(is.null(sub_lcs)) return(NULL)
-        # sub_lcs <- .perfect_coverage_simplifier_sl(sub_lcs, sub_df, t_classes_counts)
-        class(sub_lcs) <- "rlcs"
-        return(sub_lcs)
-      }
-
-      for(j in 1:length(agents)) {
-        print(paste("agent set", j, "length", length(agents[[j]]$pop)))
-        for(i in 1:length(agents[[j]]$pop)) {
-          lcs$pop[[length(lcs$pop)+1]] <- agents[[j]]$pop[[i]]
-        }
-      }
-
-      lcs$pop <- .apply_subsumption_whole_pop_sl(lcs$pop)
-      lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
-      # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
-
-      print(length(lcs$pop))
-      # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
-      # return(lcs)
-    } else {
-      for(second_evol_iter in 1:second_evolution_iterations) {
-
-        if(second_evol_iter > 1 && !is.null(second_evolution_run_params))
-          run_params <- second_evolution_run_params
-
-        print(paste("Using foreach() %dopar% to train up to", n_agents, "parallel agents."))
-        `%dopar%` <- foreach::`%dopar%`
-
-        # browser()
-        if(use_validation) {
-          validation_set <- sample(1:nrow(train_env_df), max(round(.1*nrow(train_env_df)), 1), replace = F)
-          sub_train_environment <- train_env_df[-validation_set,]
-        } else {
-          sub_train_environment <- train_env_df
-        }
-
-        agents <- foreach::foreach(i = 1:n_agents) %dopar% { ## Train N agents
-
-          sub_train_environment_shuffle <- sub_train_environment[sample(1:nrow(sub_train_environment),
-                                                                        nrow(sub_train_environment),
-                                                                        replace = F), ]
-          t_classes_counts <- table(sub_train_environment_shuffle$class) ## For Coverage!!
-          library(RLCS) ## Assuming you've gotten the package installed by now...
-
-          size_env <- nrow(sub_train_environment)
-
-          for(epoch in 1:(run_params$get_n_epochs())) {
-            for(i in 1:size_env) {
-              lcs <- .rlcs_train_one_instance_one_epoch_mat(lcs,
-                                                            sub_train_environment_shuffle[i, ],
-                                                            size_env,
-                                                            epoch,
-                                                            (epoch-1)*size_env+i, ## train_count
-                                                            run_params)
-            }
-
-
-          }
-          ## Sometimes, deletion removes all rules as none are good enough!
-          if(is.null(lcs)) return(NULL)
-          ##
-          ## Final simplification: Coverage
-          ##
-          # lcs <- .perfect_coverage_simplifier_sl(lcs, sub_train_environment_shuffle, t_classes_counts)
-
-          class(lcs) <- "rlcs"
-          return(lcs)
-        }
-        print("Now checking agents quality")
-
-        agents_quality <- list()
-        for(j in 1:n_agents) {
-          ## ADD CHECK HERE FOR !is.null(agents[[j]])
-          if(!is.null(agents[[j]])) {
-            ## Let's see how we could do testing:
-            ## We calculate accuracy BOTH for training...
-            validation_environment <- sub_train_environment[sample(1:nrow(sub_train_environment), min(1, round(0.1*nrow(train_env_df))), replace = F),]
-            validation_environment$predicted <- -1 ## Stands for not found
-            validation_environment$predicted <- rlcs_predict_sl(validation_environment, agents[[j]])
-
-            agents_quality[[j]] <- round(sum(vapply(1:nrow(validation_environment), \(i) {
-              ifelse(validation_environment[i, "class"] == validation_environment[i, "predicted"], 1, 0)
-            }, numeric(1)))/nrow(validation_environment), 4)
-
-            if(use_validation) {
-              ## AND validation:
-              validation_environment <- train_env_df[validation_set,]
-              validation_environment$predicted <- -1 ## Stands for not found
-              validation_environment$predicted <- rlcs_predict_sl(validation_environment, agents[[j]])
-
-              agents_quality[[j]] <- (agents_quality[[j]] +
-                                        round(sum(sapply(1:nrow(validation_environment), \(i) {
-                                          ifelse(validation_environment[i, "class"] == validation_environment[i, "predicted"], 1, 0)
-                                        }))/nrow(validation_environment), 4))/2
-            }
-          } else
-            agents_quality[[j]] <- 0
-        }
-
-        print(unlist(agents_quality))
-
-        if((merge_best_n > 1) & (merge_best_n <= n_agents)) {
-
-          best_agents <- order(unlist(agents_quality), decreasing = TRUE)[1:merge_best_n]
-          print(best_agents)
-
-          ## Recollect all sub-lcs
-          compacted_classifier <- list()
-          agents <- agents[best_agents]
-          # t_start_1 <- Sys.time()
-          for(j in 1:length(agents)) {
-            for(k in 1:length(agents[[j]]$pop)) {
-              compacted_classifier[[length(compacted_classifier)+1]] <- agents[[j]]$pop[[k]]
-            }
-          }
-
-          compacted_classifier <- .apply_subsumption_whole_pop_sl(compacted_classifier)
-          lcs$pop <- compacted_classifier
-          lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
-          # lcs$lengths <- sapply(lcs$pop, \(x) x$length_fixed_bits) ## Poor naming...
-          lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, integer(1)) ## Poor naming...
-          lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
-        } else {
-          best_agent <- order(unlist(agents_quality), decreasing = TRUE)[1]
-          print(best_agent)
-          lcs <- agents[[best_agent]]
-        }
-
-        # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
-      }
-    }
-
-    # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
-    # return(lcs)
-  } else {
-    print("Running single-core/thread, sequential")
-    size_env <- nrow(train_env_df)
-    ## Expose algorithm to training set:
-    for(epoch in 1:(run_params$get_n_epochs())) {
-      for(i in 1:size_env) {
-        ## Now this part of the algorithm is "necessarily" sequential...
-        lcs <- .rlcs_train_one_instance_one_epoch_mat(lcs,
-                                                      train_env_df[i, ],
-                                                      size_env,
-                                                      epoch,
-                                                      (epoch-1)*size_env+i, ## train_count
-                                                      run_params)
-        if(verbose) { ## Truly not recommended!
-          if(!is.null(lcs))
-            class(lcs) <- "rlcs"
-          message(print(lcs))
-        }
-      }
-
-    }
-    ##
-    ## Final simplification: Coverage
-    ##
-    lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
-  }
-
-  ## Sometimes, deletion removes all rules as none are good enough!
-  if(is.null(lcs$pop)) return(NULL)
-  class(lcs) <- "rlcs"
-
-  lcs
-}
+# .rlcs_train_sl_old <- function(train_env_df, run_params = RLCS_hyperparameters(),
+#                               pre_trained_lcs = NULL, verbose=FALSE,
+#                               n_agents = 0,
+#                               split_horizontal = F, ## That is but one option!
+#                               use_validation=F,
+#                               merge_best_n = 0,
+#                               second_evolution_iterations = 1,
+#                               second_evolution_run_params = NULL,
+#                               max_pop_size_parallel = 10000) {
+#   ## Initialization:
+#   lcs <- .new_rlcs()
+#   .validate_SL_train_df(train_env_df)
+#
+#   ## TODO Add Running Params Checks here...
+#   ## Further testing will surface issues here.
+#
+#   ## Re-training, or "online" updates
+#   if(!is.null(pre_trained_lcs)) lcs <- pre_trained_lcs
+#
+#   ## NEW
+#   lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+#   # lcs$lengths <- sapply(lcs$pop, \(x) x$length_fixed_bits) ## Poor naming...
+#   lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, integer(1)) ## Poor naming...
+#   lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
+#
+#   ## Shuffling population, just in case...
+#   train_env_df <- train_env_df[sample(1:nrow(train_env_df),
+#                                       nrow(train_env_df),
+#                                       replace = F), ]
+#
+#   t_classes_counts <- table(train_env_df$class) ## For Coverage!!
+#
+#   if(requireNamespace("foreach", quietly=T) &
+#      requireNamespace("doParallel", quietly=T) &
+#      n_agents > 1) { ## NEW!
+#
+#     if(split_horizontal) {
+#
+#       agents <- foreach::foreach(i = 1:n_agents) %dopar% { ## Train N agents
+#
+#         sets_size <- floor(nrow(train_env_df) / n_agents)
+#         sub_start <- (i-1)*sets_size+1
+#         sub_end <- i*sets_size
+#         sub_df <- train_env_df[sub_start:sub_end,]
+#         t_classes_counts <- table(sub_df$class) ## For Coverage!!
+#
+#         library(RLCS) ## Assuming you've gotten the package installed by now...
+#
+#         size_env <- nrow(sub_df)
+#
+#         sub_lcs <- lcs
+#
+#         for(epoch in 1:(run_params$get_n_epochs())) {
+#           for(i in 1:size_env) {
+#             sub_lcs <- .rlcs_train_one_instance_one_epoch_mat(sub_lcs,
+#                                                               sub_df[i, ],
+#                                                               size_env,
+#                                                               epoch,
+#                                                               (epoch-1)*size_env+i, ## train_count
+#                                                               run_params)
+#           }
+#         }
+#
+#         ## Sometimes, deletion removes all rules as none are good enough!
+#         if(is.null(sub_lcs)) return(NULL)
+#         # sub_lcs <- .perfect_coverage_simplifier_sl(sub_lcs, sub_df, t_classes_counts)
+#         class(sub_lcs) <- "rlcs"
+#         return(sub_lcs)
+#       }
+#
+#       for(j in 1:length(agents)) {
+#         print(paste("agent set", j, "length", length(agents[[j]]$pop)))
+#         for(i in 1:length(agents[[j]]$pop)) {
+#           lcs$pop[[length(lcs$pop)+1]] <- agents[[j]]$pop[[i]]
+#         }
+#       }
+#
+#       lcs$pop <- .apply_subsumption_whole_pop_sl(lcs$pop)
+#       lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
+#       # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
+#
+#       print(length(lcs$pop))
+#       # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
+#       # return(lcs)
+#     } else {
+#       for(second_evol_iter in 1:second_evolution_iterations) {
+#
+#         if(second_evol_iter > 1 && !is.null(second_evolution_run_params))
+#           run_params <- second_evolution_run_params
+#
+#         print(paste("Using foreach() %dopar% to train up to", n_agents, "parallel agents."))
+#         `%dopar%` <- foreach::`%dopar%`
+#
+#         # browser()
+#         if(use_validation) {
+#           validation_set <- sample(1:nrow(train_env_df), max(round(.1*nrow(train_env_df)), 1), replace = F)
+#           sub_train_environment <- train_env_df[-validation_set,]
+#         } else {
+#           sub_train_environment <- train_env_df
+#         }
+#
+#         agents <- foreach::foreach(i = 1:n_agents) %dopar% { ## Train N agents
+#
+#           sub_train_environment_shuffle <- sub_train_environment[sample(1:nrow(sub_train_environment),
+#                                                                         nrow(sub_train_environment),
+#                                                                         replace = F), ]
+#           t_classes_counts <- table(sub_train_environment_shuffle$class) ## For Coverage!!
+#           library(RLCS) ## Assuming you've gotten the package installed by now...
+#
+#           size_env <- nrow(sub_train_environment)
+#
+#           for(epoch in 1:(run_params$get_n_epochs())) {
+#             for(i in 1:size_env) {
+#               lcs <- .rlcs_train_one_instance_one_epoch_mat(lcs,
+#                                                             sub_train_environment_shuffle[i, ],
+#                                                             size_env,
+#                                                             epoch,
+#                                                             (epoch-1)*size_env+i, ## train_count
+#                                                             run_params)
+#             }
+#
+#
+#           }
+#           ## Sometimes, deletion removes all rules as none are good enough!
+#           if(is.null(lcs)) return(NULL)
+#           ##
+#           ## Final simplification: Coverage
+#           ##
+#           # lcs <- .perfect_coverage_simplifier_sl(lcs, sub_train_environment_shuffle, t_classes_counts)
+#
+#           class(lcs) <- "rlcs"
+#           return(lcs)
+#         }
+#         print("Now checking agents quality")
+#
+#         agents_quality <- list()
+#         for(j in 1:n_agents) {
+#           ## ADD CHECK HERE FOR !is.null(agents[[j]])
+#           if(!is.null(agents[[j]])) {
+#             ## Let's see how we could do testing:
+#             ## We calculate accuracy BOTH for training...
+#             validation_environment <- sub_train_environment[sample(1:nrow(sub_train_environment), min(1, round(0.1*nrow(train_env_df))), replace = F),]
+#             validation_environment$predicted <- -1 ## Stands for not found
+#             validation_environment$predicted <- rlcs_predict_sl(validation_environment, agents[[j]])
+#
+#             agents_quality[[j]] <- round(sum(vapply(1:nrow(validation_environment), \(i) {
+#               ifelse(validation_environment[i, "class"] == validation_environment[i, "predicted"], 1, 0)
+#             }, numeric(1)))/nrow(validation_environment), 4)
+#
+#             if(use_validation) {
+#               ## AND validation:
+#               validation_environment <- train_env_df[validation_set,]
+#               validation_environment$predicted <- -1 ## Stands for not found
+#               validation_environment$predicted <- rlcs_predict_sl(validation_environment, agents[[j]])
+#
+#               agents_quality[[j]] <- (agents_quality[[j]] +
+#                                         round(sum(sapply(1:nrow(validation_environment), \(i) {
+#                                           ifelse(validation_environment[i, "class"] == validation_environment[i, "predicted"], 1, 0)
+#                                         }))/nrow(validation_environment), 4))/2
+#             }
+#           } else
+#             agents_quality[[j]] <- 0
+#         }
+#
+#         print(unlist(agents_quality))
+#
+#         if((merge_best_n > 1) & (merge_best_n <= n_agents)) {
+#
+#           best_agents <- order(unlist(agents_quality), decreasing = TRUE)[1:merge_best_n]
+#           print(best_agents)
+#
+#           ## Recollect all sub-lcs
+#           compacted_classifier <- list()
+#           agents <- agents[best_agents]
+#           # t_start_1 <- Sys.time()
+#           for(j in 1:length(agents)) {
+#             for(k in 1:length(agents[[j]]$pop)) {
+#               compacted_classifier[[length(compacted_classifier)+1]] <- agents[[j]]$pop[[k]]
+#             }
+#           }
+#
+#           compacted_classifier <- .apply_subsumption_whole_pop_sl(compacted_classifier)
+#           lcs$pop <- compacted_classifier
+#           lcs$matrices <- .recalculate_pop_matrices(lcs$pop) ## Poor naming...
+#           # lcs$lengths <- sapply(lcs$pop, \(x) x$length_fixed_bits) ## Poor naming...
+#           lcs$lengths <- vapply(lcs$pop, \(x) x$length_fixed_bits, integer(1)) ## Poor naming...
+#           lcs$actions_vec <- .recalculate_actions_vec(lcs$pop)
+#         } else {
+#           best_agent <- order(unlist(agents_quality), decreasing = TRUE)[1]
+#           print(best_agent)
+#           lcs <- agents[[best_agent]]
+#         }
+#
+#         # lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
+#       }
+#     }
+#
+#     # lcs <- .apply_deletion_sl(lcs, max_pop_size = max_pop_size_parallel)
+#     # return(lcs)
+#   } else {
+#     print("Running single-core/thread, sequential")
+#     size_env <- nrow(train_env_df)
+#     ## Expose algorithm to training set:
+#     for(epoch in 1:(run_params$get_n_epochs())) {
+#       for(i in 1:size_env) {
+#         ## Now this part of the algorithm is "necessarily" sequential...
+#         lcs <- .rlcs_train_one_instance_one_epoch_mat(lcs,
+#                                                       train_env_df[i, ],
+#                                                       size_env,
+#                                                       epoch,
+#                                                       (epoch-1)*size_env+i, ## train_count
+#                                                       run_params)
+#         if(verbose) { ## Truly not recommended!
+#           if(!is.null(lcs))
+#             class(lcs) <- "rlcs"
+#           message(print(lcs))
+#         }
+#       }
+#
+#     }
+#     ##
+#     ## Final simplification: Coverage
+#     ##
+#     lcs <- .perfect_coverage_simplifier_sl(lcs, train_env_df, t_classes_counts)
+#   }
+#
+#   ## Sometimes, deletion removes all rules as none are good enough!
+#   if(is.null(lcs$pop)) return(NULL)
+#   class(lcs) <- "rlcs"
+#
+#   lcs
+# }